+++ lib/src/structures/device_auth.dart
@@ -8,9 +8,13 @@ class DeviceAuth {
   /// The secret for the device auth
   late String secret;
 
+  /// display name for user of device auth
+  late String displayName;
+
   DeviceAuth({
     required this.accountId,
     required this.deviceId,
     required this.secret,
+    this.displayName = "",
   });
 }